refactor(orc8r): ServiceRegistry service security changes
