+++ orc8r/cloud/go/services/service_registry/servicers/protected/docker_servicer.go
